break up and document nodes package
